+++ src/project/contract/vyper/expression/instruction/goto.rs
@@ -36,11 +36,11 @@ impl Goto {
     {
         let function = context
             .get_function(label_name.as_str())
-            .ok_or_else(|| anyhow::anyhow!("Function `{}` does not exist", label_name))?;
+            .ok_or_else(|| anyhow::anyhow!("Function `{label_name}` does not exist"))?;
 
         let mut arguments = Vec::new();
         for expression in self.0.into_iter() {
-            if let Expression::Identifier(ref identifier) = expression {
+            if let Ok(identifier) = expression.try_into_identifier() {
                 if identifier.starts_with(crate::r#const::LABEL_DESTINATION_PREFIX) {
                     continue;
                 }
@@ -77,7 +77,10 @@ impl Goto {
             .value
             .get_basic_blocks()
             .iter()
-            .find(|block| block.get_name().to_string_lossy() == label_name)
+            .find(|block| {
+                block.get_name().to_string_lossy()
+                    == Expression::safe_label(label_name.as_str()).as_str()
+            })
             .copied()
             .ok_or_else(|| anyhow::anyhow!("Block `{}` does not exist", label_name))?;
 
@@ -121,6 +124,6 @@ impl Goto {
             return self.into_block_call(context, label_name);
         }
 
-        self.into_function_call(context, label_name)
+        self.into_function_call(context, Expression::safe_label(label_name.as_str()))
     }
 }