+++ src/vyper/standard_json/input/mod.rs
@@ -14,6 +14,7 @@ use rayon::iter::ParallelIterator;
 use serde::Serialize;
 
 use self::language::Language;
+use self::settings::optimize::Optimize;
 use self::settings::selection::Selection;
 use self::settings::Settings;
 use self::source::Source;
@@ -41,7 +42,8 @@ impl Input {
         paths: &[PathBuf],
         evm_version: Option<era_compiler_common::EVMVersion>,
         output_selection: BTreeMap<String, Vec<Selection>>,
-        optimize: bool,
+        optimize: Optimize,
+        enable_decimals: bool,
         fallback_to_optimizing_for_size: bool,
         llvm_options: Vec<String>,
     ) -> anyhow::Result<Self> {
@@ -62,6 +64,7 @@ impl Input {
                 evm_version,
                 output_selection,
                 optimize,
+                enable_decimals,
                 fallback_to_optimizing_for_size,
                 llvm_options,
             ),
@@ -77,7 +80,8 @@ impl Input {
         sources: BTreeMap<String, String>,
         evm_version: Option<era_compiler_common::EVMVersion>,
         output_selection: BTreeMap<String, Vec<Selection>>,
-        optimize: bool,
+        optimize: Optimize,
+        enable_decimals: bool,
         fallback_to_optimizing_for_size: bool,
         llvm_options: Vec<String>,
     ) -> anyhow::Result<Self> {
@@ -93,6 +97,7 @@ impl Input {
                 evm_version,
                 output_selection,
                 optimize,
+                enable_decimals,
                 fallback_to_optimizing_for_size,
                 llvm_options,
             ),